LPS-47908 Move LiferayParser to org.pegdown because it needs to necessarily be there because of a limitation in the third party library
