+++ modules/util/markdown-converter/bnd.bnd
@@ -1,5 +1,6 @@
 Bundle-Name: Liferay Markdown Converter
 Export-Package:\
 	com.liferay.markdown.converter,\
-	com.liferay.markdown.converter.factory
+	com.liferay.markdown.converter.factory,\
+	org.pegdown.*
 Private-Package: com.liferay.markdown.converter.internal.*
